// _lib/sidebar/discover/type.ts
export type TotalCategoryType = {
  category: string[];
  total: number;
};

export type TypeCategoryTypes = {
  tar_iu_product: number;
  url: string;
  created_at: Date;
};

export type TypeCategoryDescriptionTypes = {
  tar_iu_product: number;
  description: string;
  url: string;
  hashtag: string[];
  category: string[];
  total_like: number;
  total_dislike: number;
  created_at: Date;
};